Safari-611.1.10/LayoutTests/fast/text/font-monospaced-lie-expected.txt

PASS 180 is >= document.getElementById('target').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
Hello World
